pyproject.toml
[tool.poetry]
name = "p3orm"
version = "0.1.0"
description = "Python PyPika Postgres ORM"
authors = ["Rafal Stapinski <stapinskirafal@gmail.com>"]
keywords = ["postgres", "async", "orm"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/rafalstapinski/porm"
repository = "https://github.com/rafalstapinski/porm"

[tool.poetry.dependencies]
python = "^3.10"
PyPika = "^0.48.8"
asyncpg = "^0.24.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^4.0.1"
pylint = "^2.11.1"
isort = "^5.9.3"
pytest-postgresql = "3.1.3"
psycopg2-binary = "^2.9.3"
pytest-asyncio = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_local_folder = ["test"]
